Fix a bug in finishing of heterogeneous data
There was an extremely subtle bug in bounding volume oversample in the case
of heterogeneous height maps. This case had never occurred because all height
maps were global. It emerges now due to the fact that pruning a height map
makes it heterogeneos.

Previously I made the erroneous assumption that the sorted array of real
pages is a prefix of the sorted array of real plus virtual pages. With this
assumption, the real index table is aligned with the real plus virtual offset
table. The fix involved using only the real plus virtual index table when
computing page bounds.